test: extract outbound policy coverage
(cherry picked from commit cd72fa6e77f0651df6f07b792b0ff47ea973a278)
