+++ src/infra/outbound/outbound.test.ts
@@ -24,11 +24,6 @@ import {
   formatGatewaySummary,
   formatOutboundDeliverySummary,
 } from "./format.js";
-import {
-  applyCrossContextDecoration,
-  buildCrossContextDecoration,
-  enforceCrossContextPolicy,
-} from "./outbound-policy.js";
 import { resolveOutboundSessionRoute } from "./outbound-session.js";
 import {
   formatOutboundPayloadLog,
@@ -799,63 +794,6 @@ describe("formatGatewaySummary", () => {
     for (const testCase of cases) {
       expect(formatGatewaySummary(testCase.input), testCase.name).toBe(testCase.expected);
     }
-  });
-});
-
-const slackConfig = {
-  channels: {
-    slack: {
-      botToken: "xoxb-test",
-      appToken: "xapp-test",
-    },
-  },
-} as RemoteClawConfig;
-
-const discordConfig = {
-  channels: {
-    discord: {},
-  },
-} as RemoteClawConfig;
-
-describe("outbound policy", () => {
-  it("allows cross-provider sends when enabled", () => {
-    const cfg = {
-      ...slackConfig,
-      tools: {
-        message: { crossContext: { allowAcrossProviders: true } },
-      },
-    } as RemoteClawConfig;
-
-    expect(() =>
-      enforceCrossContextPolicy({
-        cfg,
-        channel: "telegram",
-        action: "send",
-        args: { to: "telegram:@ops" },
-        toolContext: { currentChannelId: "C12345678", currentChannelProvider: "slack" },
-      }),
-    ).not.toThrow();
-  });
-
-  it("uses components when available and preferred", async () => {
-    const decoration = await buildCrossContextDecoration({
-      cfg: discordConfig,
-      channel: "discord",
-      target: "123",
-      toolContext: { currentChannelId: "C12345678", currentChannelProvider: "discord" },
-    });
-
-    expect(decoration).not.toBeNull();
-    const applied = applyCrossContextDecoration({
-      message: "hello",
-      decoration: decoration!,
-      preferComponents: true,
-    });
-
-    expect(applied.usedComponents).toBe(true);
-    expect(applied.componentsBuilder).toBeDefined();
-    expect(applied.componentsBuilder?.("hello").length).toBeGreaterThan(0);
-    expect(applied.message).toBe("hello");
   });
 });
